Create database in models instead of app
This removes the circular dependancy when trying to call models by
itself

DB is still initialized in app.py with db.init_app
