+++ models.py
@@ -1,12 +1,10 @@
-from app import db, login
+from flask_sqlalchemy import SQLAlchemy
 from flask_login import UserMixin
 from werkzeug.security import check_password_hash
 
 # From https://flask-user.readthedocs.io/en/latest/data_models.html
 
-@login.user_loader
-def load_user(id):
-    return User.query.get(int(id))
+db = SQLAlchemy()
 
 class User(UserMixin, db.Model):
     __tablename__ = "users"
@@ -43,6 +41,3 @@ class Play(db.Model):
 
     def __repr__(self):
         return f'<Play {self.word} at {self.time}>'
-
-db.create_all()
-db.session.commit()
